requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.1
langchain-groq>=0.1.0
langchain-huggingface>=0.0.1
pypdf>=3.17.0
chromadb>=0.4.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
huggingface-hub>=0.19.0
numpy>=1.24.0
pydantic>=2.0.0
typing-extensions>=4.8.0
